+++ src/components/SignInForm/SignInForm.module.css
@@ -1,0 +1,207 @@
+.signinForm {
+  background-color: #f0eff4;
+  border-radius: 30px;
+  height: 780px;
+  padding: 0 16px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+}
+@media screen and (min-width: 768px) {
+  .signinForm {
+    height: 960px;
+    padding: 0 134px;
+  }
+}
+@media screen and (min-width: 1440px) {
+  .signinForm {
+    padding: 0 118px;
+    width: 672px;
+    height: 736px;
+  }
+}
+
+
+
+.button {
+    font-family: Poppins;
+    display: flex;
+    width: 100%;
+    height: 50px;
+    justify-content: center;
+    align-items: center;
+    border-color: transparent;
+    border-radius: 30px;
+    background: #9be1a0;
+    color: #323f47;
+    font-size: 16px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: 18px;
+    letter-spacing: -0.01em;
+    margin-bottom: 16px;
+    cursor: pointer;
+  }
+
+
+
+.form {
+  width: 436px;
+  height: 398px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.eyeIcon {
+  position: absolute;
+  right: 15px;
+  bottom: 10px;
+  cursor: pointer;
+}
+
+.error {
+  color: #ef5050;
+  font-family: Poppins;
+  font-size: 14px;
+  font-weight: 400;
+  line-height: 18px;
+  letter-spacing: -0.01em;
+  text-align: left;
+  margin-top: -26px;
+}
+
+.formTitle {
+  font-family: Poppins;
+  font-size: 36px;
+  font-weight: 700;
+  line-height: 38px;
+  letter-spacing: -0.01em;
+  text-align: left;
+  margin-bottom: 25px;
+  color: #323f47;
+}
+
+.field {
+  position: relative;
+  border-radius: 15px;
+  border: 1px solid #2f2f2f;
+  outline: none;
+  display: flex;
+  width: 100%;
+  height: 50px;
+  padding-left: 16px;
+  color: #2f2f2f;
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 18px;
+  letter-spacing: -0.14px;
+  background-color: transparent;
+}
+
+
+@media screen and (min-width: 768px) {
+    .field,
+    .field::placeholder {
+      font-size: 16px;
+      line-height: 24px;
+      letter-spacing: -0.16px;
+      height: 56px;
+    }
+  }
+
+
+
+
+.field::placeholder {
+  font-family: Poppins;
+  font-size: 16px;
+  font-weight: 400;
+  line-height: 24px;
+  letter-spacing: -0.01em;
+  text-align: left;
+  color: #2f2f2f;
+  padding-left: 7px;
+}
+
+.fieldError {
+  border-color: #ef5050;
+}
+
+.fieldError::placeholder {
+  color: #ef5050;
+}
+
+.reminder {
+  margin-top: 16px;
+  text-align: center;
+  font-family: Poppins;
+  font-size: 16px;
+  font-weight: 400;
+  line-height: 24px;
+  letter-spacing: -0.01em;
+  color: #2f2f2f;
+}
+
+.link {
+  font-family: Poppins;
+  font-size: 16px;
+  font-weight: 700;
+  line-height: 24px;
+  letter-spacing: -0.01em;
+  text-align: left;
+  text-decoration: underline;
+  color: #2f2f2f;
+}
+
+.logo {
+  position: absolute;
+  z-index: 999;
+  top: 24px;
+  left: 16px;
+  color: var(--color-darkblue);
+  font-size: 24px;
+  font-weight: 700;
+  line-height: 1;
+  letter-spacing: -0.24px;
+  text-transform: uppercase;
+}
+@media screen and (min-width: 768px) {
+  .logo {
+    top: 32px;
+    left: 32px;
+  }
+}
+
+.signInLabel {
+  color: #2f2f2f;
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 18px;
+  letter-spacing: -0.14px;
+}
+@media screen and (min-width: 768px) {
+  .signInLabel {
+    font-size: 16px;
+    line-height: 24px;
+    letter-spacing: -0.16px;
+  }
+}
+.signinFormGroupEmail {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  margin-bottom: 20px;
+  width: 100%;
+}
+.signinFormGroupPassword {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  margin-bottom: 32px;
+  width: 100%;
+  position: relative;
+}
